+++ components/plug-socket/src/plug.js
@@ -1,0 +1,6 @@
+AFRAME.registerComponent('plug', {
+
+  init() {
+    this.el.setAttribute('socket', {type: 'plug'})
+  }
+})
