+++ components/plug-socket/test/unit-tests.html
@@ -9,7 +9,9 @@
 <body>
   <div id="qunit"></div>
   <div id="qunit-fixture"></div>
-  <script src="../dist/plug-socket.js"></script>
-  <script src = "./ut/add.js"></script>
+  <script src="../src/socket.js"></script>
+  <script src="../src/socket-fabric.js"></script>
+  <script src="../src/socket-system.js"></script>
+  <script src="../src/plug.js"></script>
   <script src = "./ut/ut1.js"></script>
 </body>
