--- components/plug-socket/test/ut/add.js
@@ -1,3 +1,0 @@
-function add(a, b) {
-  return a + b;
-}
